requirements.txt
pyserial==3.5
openai==1.3.0
requests==2.31.0